@@ -1,8 +1,8 @@
 <script>
-	import { auth } from "$lib/firebase";
-	import { signOut } from "firebase/auth";
+	import { auth } from '$lib/firebase';
+	import { signOut } from 'firebase/auth';
 
-	let isNavOpen = true;
+	let isNavOpen = $state(true);
 
 	function toggleNav() {
 		isNavOpen = !isNavOpen;
@@ -15,13 +15,13 @@
 
 {#if isNavOpen}
 	<nav
-		class=" top-0 left-0 h-screen w-64 bg-secondary text-primary flex flex-col justify-between items-center"
+		class=" left-0 top-0 flex h-screen w-64 flex-col items-center justify-between bg-secondary text-primary"
 	>
-		<div class="w-full flex flex-col justify-center items-center relative">
-			<button class="absolute top-0 left-0 text-lg w-12 h-12" on:click={toggleNav}>
+		<div class="relative flex w-full flex-col items-center justify-center">
+			<button class="absolute left-0 top-0 h-12 w-12 text-lg" onclick={toggleNav}>
 				<i class="bi bi-x"></i>
 			</button>
-			<h2 class="text-2xl font-bold my-10 p-2">PennyWise</h2>
+			<h2 class="my-10 p-2 text-2xl font-bold">PennyWise</h2>
 
 			<ul class="p-2">
 				<li class="mb-2"><a href="/overview" class="text-lg hover:text-gray-400">Overview</a></li>
@@ -33,10 +33,10 @@
 				<li class="mb-2"><a href="/Recurring" class="text-lg hover:text-gray-400">Recurring</a></li>
 			</ul>
 		</div>
-		<button class="text-lg hover:text-gray-400 p-4" on:click={signOutSSR}>Sign Out</button>
+		<button class="p-4 text-lg hover:text-gray-400" onclick={signOutSSR}>Sign Out</button>
 	</nav>
 {:else}
-	<button class="top-0 left-0 m-6 h-full text-xl" on:click={toggleNav}>
+	<button class="left-0 top-0 m-6 h-full text-xl" onclick={toggleNav}>
 		<i class="bi bi-list"></i>
 	</button>
 {/if}
