@@ -1,8 +1,8 @@
-import { user } from "$lib/stores/firebaseUser";
-import type { Load } from "@sveltejs/kit";
+import { user } from '$lib/stores/firebaseUser';
+import type { Load } from '@sveltejs/kit';
 
 export const load: Load = async () => {
-  return {
-    user: user
-  };
+	return {
+		user: user
+	};
 };
