@@ -12,13 +12,13 @@ export const load = (async ({ locals }) => {
 	const transactions = transactionDoc.docs.map((doc) => {
 		const data = doc.data();
 		return {
-				...data,
-				createdAt: data.createdAt.toDate().toISOString()
+			...data,
+			dueDate: data.dueDate?.toDate().toISOString(),
+			createdAt: data.createdAt.toDate().toISOString()
 		};
-});
+	});
 	const userData = userDoc.data();
 	if (!userData) throw error(404, 'User not found');
-	console.log(transactions);
 
 	return {
 		user: userData,
@@ -37,13 +37,15 @@ export const actions = {
 		const price = formData.get('price');
 		const monthly = formData.get('monthly') === 'on';
 		const category = formData.get('category');
+		const dueDate = formData.get('dueDate');
 
 		const newItem = {
 			name,
 			price,
 			monthly,
 			category,
-			createdAt: Timestamp.now()
+			createdAt: Timestamp.now(),
+			dueDate: dueDate ? Timestamp.fromDate(new Date(dueDate.toString())) : null
 		};
 		await adminDB.collection(`users/${uid}/transactions`).add(newItem);
 
